move error folder
