+++ apps/server/src/modules/files-storage/api/controller/api-test/files-storage-preview.api.spec.ts
@@ -14,7 +14,7 @@ import { TestApiClient } from '@testing/test-api-client';
 import NodeClam from 'clamscan';
 import FileType from 'file-type-cjs/file-type-cjs-index';
 import { FileRecord, ScanStatus } from '../../../repo/entity';
-import { ErrorType } from '../../../error';
+import { ErrorType } from '../../../domain/error';
 import { FilesStorageTestModule } from '../../../files-storage-test.module';
 import { FILES_STORAGE_S3_CONNECTION } from '../../../files-storage.config';
 import { TestHelper } from '../../../domain/helper/test-helper';
