+++ apps/server/src/modules/files-storage/domain/error/index.ts
