+++ apps/server/src/modules/files-storage/domain/index.ts
@@ -1,3 +1,4 @@
 export { PreviewService, FilesStorageService } from './service';
 export { FilesStorageConfigResponse, FileDto } from './dto';
 export * from './interface';
+export { ErrorType } from './error';
