+++ apps/server/src/modules/files-storage/domain/service/files-storage-download.service.spec.ts
@@ -9,7 +9,7 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { setupEntities } from '@testing/database';
 import { FileRecordParams } from '../../api/dto'; // TODO: invalid import
 import { FileRecord, ScanStatus } from '../../repo/entity'; // TODO: invalid import
-import { ErrorType } from '../../error';
+import { ErrorType } from '../error';
 import { FILES_STORAGE_S3_CONNECTION } from '../../files-storage.config';
 import { createPath } from '../helper';
 import { FileRecordParentType, StorageLocation } from '../interface';
