+++ apps/server/src/modules/files-storage/domain/service/files-storage-upload.service.spec.ts
@@ -12,7 +12,7 @@ import { PassThrough, Readable } from 'stream';
 import { FileRecordParams } from '../../api/dto'; // TODO: invalid import
 import { FileDto } from '../dto';
 import { FileRecord } from '../../repo/entity';
-import { ErrorType } from '../../error';
+import { ErrorType } from '../error';
 import { FILES_STORAGE_S3_CONNECTION } from '../../files-storage.config';
 import { createFileRecord, resolveFileNameDuplicates } from '../helper';
 import { FileRecordParentType, StorageLocation } from '../interface';
